requirements.txt
asgiref==3.5.2
crispy-bootstrap5==0.6
Django==4.1
django-crispy-forms==1.14.0
django-filter==22.1
djangorestframework==3.13.1
Markdown==3.4.1
pytz==2022.2.1
sqlparse==0.4.2
